+++ lib/src/decorators/not.dart
@@ -3,10 +3,13 @@ part of nodes;
 /// The [Not] decorator returns the opposite [Status] of its decorated [Node]
 /// instance.
 ///
-/// - On `Status.success`, the [Not] decorator instance returns `Status.failure`.
-/// - On `Status.failure`, the [Not] decorator instance returns `Status.success`.
-/// - On `Status.running`, the [Not] decorator instance returns `Status.running`
-/// 
+/// - On `Status.success`, the [Not] decorator instance returns
+/// `Status.failure`.
+/// - On `Status.failure`, the [Not] decorator instance returns
+/// `Status.success`.
+/// - On `Status.running`, the [Not] decorator instance returns
+/// `Status.running`
+///
 /// This is the equivalent of a logical NOT on a boolean expression.
 class Not extends Decorator {
   /// Construct a [Not] decorator for the specified argument [Node].
@@ -17,6 +20,9 @@ class Not extends Decorator {
   /// status as an actual result.
   const Not(final Node node) : super(node);
 
+  /// Updates the decorated [Node] and returns the opposite [Status]. It
+  /// inverses `Status.success` into `Status.failure` and vice-versa. If the
+  /// decorated [Node] returns `Status.running`, [Not] returns `Status.running`.
   @override
   Status update() {
     final status = _node.update();
@@ -24,7 +30,7 @@ class Not extends Decorator {
     if (status == Status.running) {
       return Status.running;
     }
-
+    // inverse Status.success into Status.failure and vice-versa
     return status == Status.success ? Status.failure : Status.success;
   }
 }