+++ lib/src/notifiers/data_node_observer.dart
@@ -1,0 +1,17 @@
+part of nodes;
+
+class DataNodeObserver<T, U> extends Observer {
+  final Node comparer;
+
+  DataNodeObserver(this.comparer);
+
+  @override
+  void receive(DataSubject<U> subject) {
+    if (comparer.update() == Status.success) {}
+  }
+
+  @override
+  Status update() {
+    return Status.success;
+  }
+}