+++ lib/src/states/state_machine.dart
@@ -1,33 +1,8 @@
 part of nodes;
 
-class _AnyStateMachine extends Node {
-  static final _invalid = _AnyStateMachine._empty();
-
-  final List<_State> _states;
-  int _index;
-  int _nextIndex;
-
-  _AnyStateMachine._empty()
-      : _states = List.empty(growable: false),
-        _index = -1,
-        _nextIndex = -1;
-
-  _AnyStateMachine(final int capacity)
-      : assert(capacity > 0),
-        _states = List.filled(capacity, _State._invalid, growable: false),
-        _index = 0,
-        _nextIndex = 0;
-
-  @override
-  Status update() {
-    assert(_states.every((state) => !identical(state, _State._invalid)));
-    return _states[_index].update();
-  }
-}
-
 /// A [StateMachine] handles the update of a state and transitions between
 /// a state and other defined states.
-class StateMachine<TEnum> extends _AnyStateMachine {
+class StateMachine<TEnum> extends _StateMachineBase {
   final List<TEnum> _enum;
 
   /// Constructs a [StateMachine] instance with the specified list of states.
@@ -51,10 +26,12 @@ class StateMachine<TEnum> extends _AnyStateMachine {
   /// - `state` is the state being defined.
   /// - `update` is a [Node] instance that is updated every time the state machine
   /// is updated.
-  /// - `enter` is a [Node] instance that is updated once when the state is first
-  /// entered.
+  /// - `enter` is a [Node] instance that is updated once when the state is
+  /// first entered.
   /// - `exit` is a [Node] instance that is updated once when the state is first
   /// exited.
+  /// - `isPartial` determines whether the state is evaluated partially, or not,
+  /// as described in the documentation of [Sequence].
   ///
   /// A state will execute its `enter` [Node] on every `update()` call
   /// to the [StateMachine] instance until `Status.success` is returned
@@ -67,24 +44,26 @@ class StateMachine<TEnum> extends _AnyStateMachine {
   ///
   /// A state will execute its `exit` [Node] if a state transition is triggered
   /// by the state or by a call to the `set(state)` method of the [StateMachine]
-  /// instance. In exit mode, every `update()` call to the [StateMachine] instance
-  /// until `Status.success` is returned by the `exit` [Node] or a transition is
-  /// triggered manually.
+  /// instance. In exit mode, every `update()` call to the [StateMachine]
+  /// instance until `Status.success` is returned by the `exit` [Node] or a
+  /// transition is triggered manually.
   void define(
     final TEnum state, {
-    final Node? update,
     final Node? enter,
+    final Node? update,
     final Node? exit,
+    final bool isPartial = false,
   }) {
-    _states[_enumAsInt(state)] =
-        _State(this, update: update, enter: enter, exit: exit);
+    _states[_enumAsInt(state)] = _State(this,
+        update: update, enter: enter, exit: exit, isPartial: isPartial);
   }
 
   /// Defines a transition from a state to another state based on a
   /// condition [Node] instance.
   ///
   /// - `from` is the state that evaluates the `on` [Node] condition.
-  /// - `to` is the state that will be transitioned to if the condition is fulfilled.
+  /// - `to` is the state that will be transitioned to if the condition is
+  /// fulfilled.
   /// - `on` is the condition [Node] that determines whether to switch states; a
   /// condition is fulfilled if the `on` [Node] returns [Status.success] from
   /// its `update()` method.