+++ lib/src/states/state_notification_machine.dart
@@ -1,0 +1,5 @@
+part of nodes;
+
+class StateNotificationMachine {
+  
+}