+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.0.4
+
+- Removed generic project example; will add real world examples later
+- Removed unused dart:ffi import
+- Improved examples of DataObserver and AutoDataObserver
+- Simplified interface of DataObserver and AutoDataObserver
+- BREAKING CHANGE: replaced named optional parameter 'action' as a formal
+required positional parameter in the makeClosure() function
+
 0.0.3
 
 - Added the subject-observer types